Move AHK proxy to seperate file
